+++ dependency/core/applications/pedestrian_cyclist_intent_recognition/__init__.py
@@ -1,3 +1,0 @@
-from .pedestrian_cyclist_intent_recognition import PedestrianCyclistIntentRecognition as Structured_Processor
-
-__all__ = ["Structured_Processor"]