+++ dependency/core/applications/pedestrian_cyclist_intent_recognition/pedestrian_cyclist_intent_recognition_with_tensorrt/pedestrian_cyclist_intent_recognition_tensorrt8.py
@@ -1,6 +1,0 @@
-class PedestrianCyclistIntentRecognitionTensorRT8:
-    def __init__(self, weights='', plugin_library='', device=0):
-        self.weights = weights
-        self.plugin_library = plugin_library
-        self.device = device
-        raise NotImplementedError('pedestrian-cyclist-intent-recognition TensorRT implementation is not implemented yet.')