+++ dependency/core/applications/pedestrian_cyclist_intent_recognition/pedestrian_cyclist_intent_recognition_without_tensorrt/__init__.py
@@ -1,3 +1,0 @@
-from .pedestrian_cyclist_intent_recognition_gru import PedestrianCyclistIntentRecognition as PedestrianCyclistIntentRecognitionWithoutTensorRT
-
-__all__ = ['PedestrianCyclistIntentRecognitionWithoutTensorRT']