+++ dependency/core/applications/pedestrian_cyclist_pose_estimation/__init__.py
@@ -1,3 +1,0 @@
-from .pedestrian_cyclist_pose_estimation import PedestrianCyclistPoseEstimation as Structured_Processor
-
-__all__ = ["Structured_Processor"]