+++ dependency/core/applications/pedestrian_cyclist_pose_estimation/pedestrian_cyclist_pose_estimation_without_tensorrt/__init__.py
@@ -1,3 +1,0 @@
-from .pedestrian_cyclist_pose_estimation_mmpose import PedestrianCyclistPoseEstimation as PedestrianCyclistPoseEstimationWithoutTensorRT
-
-__all__ = ['PedestrianCyclistPoseEstimationWithoutTensorRT']