+++ dependency/core/applications/pedestrian_pose_estimation/pedestrian_pose_estimation.py
@@ -1,0 +1,45 @@
+from core.lib.common import Context, LOGGER
+
+
+class PedestrianPoseEstimation:
+    def __init__(self, trt_weights='', trt_plugin_library='', non_trt_weights='', device=0, non_trt_config=''):
+        use_tensorrt = Context.get_parameter('USE_TENSORRT', default=False, direct=False)
+        self.trt_weights = Context.get_file_path(trt_weights) if trt_weights else ''
+        self.trt_plugin_library = Context.get_file_path(trt_plugin_library) if trt_plugin_library else ''
+        self.non_trt_weights = Context.get_file_path(non_trt_weights) if non_trt_weights else ''
+        self.device = device
+        self.non_trt_config = Context.get_file_path(non_trt_config) if non_trt_config else ''
+
+        if use_tensorrt:
+            jetpack_version = Context.get_parameter('JETPACK', direct=False)
+            if jetpack_version == 6:
+                LOGGER.info('Using TensorRT 10 (JetPack 6)')
+                from .pedestrian_pose_estimation_with_tensorrt import PedestrianPoseEstimationTensorRT10
+                self.model = PedestrianPoseEstimationTensorRT10(
+                    weights=self.trt_weights,
+                    plugin_library=self.trt_plugin_library,
+                    device=self.device,
+                )
+            else:
+                if jetpack_version not in [4, 5]:
+                    LOGGER.warning(f'Unknown JETPACK version: {jetpack_version}, attempting to use TensorRT 8')
+                else:
+                    LOGGER.info(f'Using TensorRT 8 (JetPack {jetpack_version})')
+                from .pedestrian_pose_estimation_with_tensorrt import PedestrianPoseEstimationTensorRT8
+                self.model = PedestrianPoseEstimationTensorRT8(
+                    weights=self.trt_weights,
+                    plugin_library=self.trt_plugin_library,
+                    device=self.device,
+                )
+        else:
+            from .pedestrian_pose_estimation_without_tensorrt import PedestrianPoseEstimationWithoutTensorRT
+            self.model = PedestrianPoseEstimationWithoutTensorRT(
+                weights=self.non_trt_weights,
+                device=self.device,
+                model_config=self.non_trt_config,
+            )
+
+        self.flops = getattr(self.model, 'flops', 0)
+
+    def __call__(self, payload):
+        return self.model(payload)