+++ dependency/core/applications/risk_graph_generation/risk_graph_generation_with_tensorrt/__init__.py
@@ -1,0 +1,4 @@
+from .risk_graph_generation_tensorrt8 import RiskGraphGenerationTensorRT8
+from .risk_graph_generation_tensorrt10 import RiskGraphGenerationTensorRT10
+
+__all__ = ['RiskGraphGenerationTensorRT8', 'RiskGraphGenerationTensorRT10']