+++ dependency/core/applications/risk_graph_generation/risk_graph_generation_without_tensorrt/__init__.py
@@ -1,0 +1,3 @@
+from .risk_graph_generation_mlp import RiskGraphGeneration as RiskGraphGenerationWithoutTensorRT
+
+__all__ = ['RiskGraphGenerationWithoutTensorRT']