+++ dependency/core/applications/traffic_detection/traffic_detection_without_tensorrt/__init__.py
@@ -1,0 +1,3 @@
+from .traffic_detection_yolo import TrafficDetection as TrafficDetectionWithoutTensorRT
+
+__all__ = ['TrafficDetectionWithoutTensorRT']