+++ dependency/core/applications/traffic_object_detection/__init__.py
@@ -1,3 +1,0 @@
-from .traffic_object_detection import TrafficObjectDetection as Structured_Processor
-
-__all__ = ["Structured_Processor"]