+++ dependency/core/applications/traffic_object_detection/traffic_object_detection_without_tensorrt/__init__.py
@@ -1,3 +1,0 @@
-from .traffic_object_detection_yolo import TrafficObjectDetection as TrafficObjectDetectionWithoutTensorRT
-
-__all__ = ['TrafficObjectDetectionWithoutTensorRT']