+++ dependency/core/applications/traffic_risk_graph_inference/__init__.py
@@ -1,3 +1,0 @@
-from .traffic_risk_graph_inference import TrafficRiskGraphInference as Structured_Processor
-
-__all__ = ["Structured_Processor"]