+++ dependency/core/applications/traffic_risk_graph_inference/traffic_risk_graph_inference_without_tensorrt/__init__.py
@@ -1,3 +1,0 @@
-from .traffic_risk_graph_inference_mlp import TrafficRiskGraphInference as TrafficRiskGraphInferenceWithoutTensorRT
-
-__all__ = ['TrafficRiskGraphInferenceWithoutTensorRT']