+++ dependency/core/applications/vehicle_reidentification_tracking/__init__.py
@@ -1,3 +1,0 @@
-from .vehicle_reidentification_tracking import VehicleReidentificationTracking as Structured_Processor
-
-__all__ = ["Structured_Processor"]