+++ dependency/core/applications/vehicle_reidentification_tracking/vehicle_reidentification_tracking_with_tensorrt/__init__.py
@@ -1,4 +1,0 @@
-from .vehicle_reidentification_tracking_tensorrt8 import VehicleReidentificationTrackingTensorRT8
-from .vehicle_reidentification_tracking_tensorrt10 import VehicleReidentificationTrackingTensorRT10
-
-__all__ = ['VehicleReidentificationTrackingTensorRT8', 'VehicleReidentificationTrackingTensorRT10']