+++ dependency/core/applications/vehicle_tracking/vehicle_tracking_with_tensorrt/__init__.py
@@ -1,0 +1,4 @@
+from .vehicle_tracking_tensorrt8 import VehicleTrackingTensorRT8
+from .vehicle_tracking_tensorrt10 import VehicleTrackingTensorRT10
+
+__all__ = ['VehicleTrackingTensorRT8', 'VehicleTrackingTensorRT10']