+++ CHANGELOG.md
@@ -9,7 +9,7 @@
 ### Bug Fix
 
 ### Minor Update
-- Add more services: traffic-object-detection, road-context-segmentation, traffic-signal-recognition, vehicle-reidentification-tracking, vehicle-attribute-recognition, vehicle-trajectory-prediction, pedestrian-cyclist-pose-estimation, pedestrian-cyclist-intent-recognition, traffic-risk-graph-inference (`processor`).
+- Add more services: traffic-detection, road-context-segmentation, traffic-signal-recognition, vehicle-tracking, vehicle-attribute-recognition, vehicle-trajectory-prediction, pedestrian-pose-estimation, pedestrian-intent-recognition, risk-graph-generation (`processor`).
 - Change input/output format of processors from value to list (`processor`).
 - Add import/export of dag in DAG Orchestrain frontend page (`frontend`).
 - Set default cloud processor as options to permit edge-only processors (`backend`).